Extracts types into their own file and imports them where needed
