+++ src/constants/experience.ts
@@ -17,5 +17,3 @@ export const JOB_CONTENT = {
   netDeveloperComknow: 'NetDeveloperComknow',
   iosApprenticeReea: 'IosApprenticeReea',
 } as const
-
-export type JobContentType = typeof JOB_CONTENT[keyof typeof JOB_CONTENT]
